refactor: update auth
- create client logout
- create server logout
- logout if it is not possible to get the id from the token
