+++ src/app/api/goals/change-name/route.ts
@@ -1,8 +1,11 @@
+import { NextResponse } from 'next/server';
+import { getTranslations } from 'next-intl/server';
+
 import { goalsController } from '@/controllers/goals-controller';
 import { yearsController } from '@/controllers/years-controller';
+
+import { logout } from '@/services/auth-service';
 import { checkUserIdService } from '@/services/check-user-id-service';
-import { getTranslations } from 'next-intl/server';
-import { NextResponse } from 'next/server';
 
 export async function PUT(req: Request) {
   const checkUserId = await checkUserIdService(req);
@@ -10,7 +13,7 @@ export async function PUT(req: Request) {
   const t = await getTranslations('Errors');
 
   if (!checkUserId.success) {
-    return NextResponse.json({ message: checkUserId.error }, { status: 500 });
+    return await logout();
   }
 
   const userId = checkUserId.userId;
