+++ src/components/feature/logoutActions/components/LogoutActions.tsx
@@ -1,14 +1,11 @@
 'use client';
 
 import type { FC } from 'react';
-import toast from 'react-hot-toast';
 import { useTranslations } from 'next-intl';
 
-import { PATHS } from '@/constants/paths';
-
-import { httpLogout } from '@/lib/http/auth';
-
 import BaseIcon, { ExitIcon } from '@/components/ui/icon';
+
+import { logoutClient } from '@/utils/logout';
 
 import classes from '@/components/ui/popover/styles/popover.module.css';
 
@@ -16,21 +13,13 @@ const LogoutActions: FC = () => {
   const t = useTranslations('Auth');
   const tErrors = useTranslations('Errors');
 
-  const logout = async () => {
-    const res = await httpLogout();
-
-    if (!res.ok) {
-      toast.error(tErrors('something'));
-      return;
-    }
-
-    localStorage.removeItem('latestCSRFToken');
-    window.location.href = PATHS.auth.signIn;
-  };
-
   return (
     <>
-      <button type="button" className={classes.content_action} onClick={() => logout()}>
+      <button
+        type="button"
+        className={classes.content_action}
+        onClick={async () => await logoutClient(tErrors('something'))}
+      >
         <BaseIcon>
           <ExitIcon />
         </BaseIcon>
