+++ src/components/ui/aside/hooks/useAside.ts
@@ -1,9 +1,13 @@
 'use client';
 
-import { LOCAL_STORAGE_YEARS_KEY } from '@/constants/localstorage';
-import { httpGetYears } from '@/lib/http/years';
 import { useCallback, useState } from 'react';
 import toast from 'react-hot-toast';
+
+import { LOCAL_STORAGE_YEARS_KEY } from '@/constants/localstorage';
+
+import { httpGetYears } from '@/lib/http/years';
+
+import { logoutClient } from '@/utils/logout';
 
 const useAside = (defaultYears: number[]) => {
   const [isLoading, setIsLoading] = useState(true);
@@ -26,6 +30,7 @@ const useAside = (defaultYears: number[]) => {
       updateYears(uniqueSortedYears);
     } catch (error) {
       toast.error('Failed to fetch years');
+      await logoutClient('Error logout');
       console.error('Failed to fetch years:', error);
     } finally {
       setIsLoading(false);
